requirements.txt
aniso8601==7.0.0
asgiref==3.2.3
Django==3.0.4
django-filter==2.2.0
django-js-asset==1.2.2
django-mptt==0.11.0
djangorestframework==3.11.0
graphene==2.1.8
graphene-django==2.8.2
graphene-django-optimizer==0.6.2
graphql-core==2.3.1
graphql-relay==2.0.1
Pillow==7.0.0
pkg-resources==0.0.0
promise==2.3
psycopg2==2.8.4
pytz==2019.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.1
